+++ Example/Tests/Models/TestObject.swift
@@ -16,7 +16,7 @@ class TestObject: Object, PredicateSchema {
     @objc dynamic var int: Int = 0
     @objc dynamic var date: Date?
     let children = List<TestObject>()
-    let parent = LinkingObjects<TestObject>(fromType: TestObject.self, property: "children")
+    let parent = LinkingObjects<TestObject>(TestObject.self, TestObjectSchema.children)
     
     convenience init(key: Int, string: String, int: Int, date: Date? = nil) {
         self.init()
@@ -30,13 +30,4 @@ class TestObject: Object, PredicateSchema {
     override class func primaryKey() -> String? {
         return TestObjectSchema.key.keyPath()
     }
-}
-
-extension TestObjectSchema {
-    
-    var children: CollectionProperty<TestObjectSchema> { return builder.collection("children") }
-    static var children: CollectionProperty<TestObjectSchema> { return TestObjectSchema().children }
-    
-    var parent: CollectionProperty<TestObjectSchema> { return builder.collection("parent") }
-    static var parent: CollectionProperty<TestObjectSchema> { return TestObjectSchema().parent }
 }